pyproject.toml
[tool.poetry]
name = "ssp_flask"
version = "0.1.0"
description = ""
authors = ["Tom Camp <tomcamp@protonmail.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
pydantic = "^2.8.2"
pyyaml = "^6.0.2"
flask-wtf = "^1.2.1"
rtyaml = "^1.0.0"
pytest = "^8.3.2"
coverage = "^7.6.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
